!!! TASK: Clean up Data Provider class structure
This change renames the class "DataProvider" into "AbstractDataProvider" and extracts the
database-specific code into a new class "AbstractDatabaseDataProvider". If you previously
extended "DataProvider", you should now extend "AbstractDatabaseDataProvider" instead.

Also contains further doc comments and small clean ups.
