+++ Classes/Ttree/ContentRepositoryImporter/Domain/Model/PresetPartDefinition.php
@@ -1,9 +1,9 @@
 <?php
 namespace Ttree\ContentRepositoryImporter\Domain\Model;
 
-/*                                                                                  *
- * This script belongs to the TYPO3 Flow package "Ttree.ContentRepositoryImporter". *
- *                                                                                  */
+/*
+ * This script belongs to the Neos Flow package "Ttree.ContentRepositoryImporter".
+ */
 
 use Ttree\ContentRepositoryImporter\Exception\InvalidArgumentException;
 use TYPO3\Flow\Annotations as Flow;
