+++ Classes/Ttree/ContentRepositoryImporter/Domain/Service/ImportService.php
@@ -1,9 +1,9 @@
 <?php
 namespace Ttree\ContentRepositoryImporter\Domain\Service;
 
-/*                                                                                  *
- * This script belongs to the TYPO3 Flow package "Ttree.ContentRepositoryImporter". *
- *                                                                                  */
+/*
+ * This script belongs to the Neos Flow package "Ttree.ContentRepositoryImporter".
+ */
 
 use Doctrine\ORM\Mapping as ORM;
 use Ttree\ContentRepositoryImporter\Domain\Model\Event;
